refactor: main의 로직을 BlackJackRun 클래스로 이동
