+++ src/main/java/blackjack/view/OutputView.java
@@ -7,10 +7,8 @@ import blackjack.domain.card.Card;
 import blackjack.domain.player.Dealer;
 import blackjack.domain.player.Gamer;
 import blackjack.domain.player.Player;
-import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 
 public class OutputView {
